backend/ai-service/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.2
google-generativeai
python-dotenv
tensorflow
pandas==1.0.0
